refactor(investment): 독립 페이지 → 대시보드 탭으로 통합
# 사용자 요청
"독립 페이지는 없애고, 좌측 탭 메뉴에 주식 자동 매매를 클릭했을 때 보이는
페이지로 통합해줘"

# 통합 방식
- 모든 /investment/* 메인 페이지(connect/strategy/daytrading/compare/screener/
  smart-money/trading/psychology/portfolio/rankings/rl-models, /investment 자체)
  를 /dashboard?tab=investment&sub=X 로 client redirect.
- InvestmentTab(대시보드의 'investment' 탭 콘텐츠)이 single source of truth.
- ?sub= URL 파라미터로 sub-tab 선택 동기화 (새로고침/딥링크 유지).

# InvestmentTab 변경
- SUB_TABS 에 rankings(Trophy), rl-models(Cpu) 추가
- useSearchParams 로 ?sub= 읽고 쓰기, 외부 변경 시 자동 동기화
- setSubTab 시 router.replace 로 URL 갱신 (scroll: false)

# 신규 컴포넌트
- src/components/Investment/RankingsContent.tsx: 기존 /investment/rankings 페이지 콘텐츠
- src/components/Investment/RLModelsContent.tsx: 기존 /investment/rl-models 페이지 콘텐츠 (min-h-screen 제거)

# /investment/* 메인 페이지
- 모두 단순 client redirect 컴포넌트로 변경
- 동적 라우트(strategy/[id]/{edit,backtest}, strategy/{new,preset/[id]}), 게이트
  페이지(subscribe), 별도 도메인(auction)은 그대로 유지

# 사이드바
- /investment/layout.tsx 의 NAV_ITEMS href 를 /dashboard?tab=investment&sub=X 로
  변경하여 redirect 단계를 건너뛰고 바로 통합 페이지로 진입

# 효과
- 기능 추가/수정 시 한 군데(InvestmentTab + 해당 Content 컴포넌트)만 수정하면 됨
- 사용자가 헷갈리던 "두 군데서 다르게 보이는" 현상 해결
- 기존 /investment/X 직접 링크/북마크는 redirect 로 자동 흡수
